Makes the Dice attribute of the Ability class not a list, because it doesn't need to be one.
Makes DiceCollections and DicePiles able to be deserialized.

Adds a DisplayTier attribute to the Ability class, to influence sorting.

Sets up sorting for the abilities list. The abilities list will now be sorted initially, based on DisplayTier + Name.

Hooks up the +1 and -1 buttons on the abilities list.

Sorts the functions in Form1.cs into regions based on which tab they affect.