Makes the Character class recalculate proficiency bonus periodically, because it was getting stuck before.
Hooks the character saving into the Settings class.

Breaks the Settings class out into it's own project, so it can be included in other stuff without having to worry about cyclic includes.

Adds the ability to load a character through a dropdown on the settings page, or to load a new blank character.